+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update `criterion` dev-dependency to 0.5
+
 ## [0.16.0] - 2023-10-11
 
 ### Added