+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update `criterion` dev-dependency to 0.5
+- Fix clippy warnings [#774]
 
 ## [0.16.0] - 2023-10-11
 
@@ -515,6 +516,7 @@ is necessary since `rkyv/validation` was required as a bound.
 - Proof system module.
 
 <!-- ISSUES -->
+[#774]: https://github.com/dusk-network/plonk/issues/774
 [#763]: https://github.com/dusk-network/plonk/issues/763
 [#760]: https://github.com/dusk-network/plonk/issues/760
 [#752]: https://github.com/dusk-network/plonk/pull/752